requirements.txt
fastapi[all]
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
psycopg2-binary
pymongo
requests
python-dotenv
Jinja2==3.1.2